requirements.txt
streamlit
google-generativeai
numpy
pandas
scikit-learn
plotly
python-dotenv